update endpoint for pdf
