+++ src/actions/index.js
@@ -1,10 +1,11 @@
 import { loadFeaturesFromRoute, loadFeatureFromCoords, loadFlaggedData } from "../utils/tilequery";
-import { getRegionFromFeatureId } from "../modules/config";
+import { getRegionFromFeatureId, getRegionFromFeature } from "../modules/config";
 import {FlyToInterpolator} from 'react-map-gl';
 import * as ease from 'd3-ease';
 import { addFeatureToRoute, removeFeatureFromRoute, updateRoute } from '../modules/router';
 import { getStateViewport } from '../constants/statesFips';
 import { getFeatureProperty } from "../modules/features";
+import axios from 'axios';
 
 /** ACTIONS */
 
@@ -269,10 +270,7 @@ export const onMapLegendAction = (itemId) => ({
   itemId
 })
 
-export const onReportDownload = (feature) => ({
-  type: 'REPORT_DOWNLOAD',
-  feature
-})
+
 
 export const onShowSimilar = (feature) => ({
   type: 'SHOW_SIMILAR',
@@ -493,3 +491,39 @@ export const toggleHelp = (forceOpen = false) =>
       open: !helpOpen || forceOpen
     })
   }
+
+  const getPdfRegion = (region) => {
+    return region === 'counties' ? 'county' :
+      (region === 'districts') ? 'district' : 'school'
+  }
+
+  export const onReportDownload = (feature) => dispatch => {
+    dispatch({
+      type: 'REPORT_DOWNLOAD',
+      feature
+    })
+    const region = getRegionFromFeature(feature);
+    axios({
+      url: 'http://export.edopportunity.org/',
+      method: 'POST',
+      data: {
+        "region": getPdfRegion(region),
+        "location": {
+          ...feature.properties,
+          "diff_avg": -1.244,
+          "diff_grd": -12,
+          "diff_coh": 1.34499
+        },
+        "url": window.location.href,
+        "others": null
+      },
+      responseType: 'blob', // important
+    }).then((response) => {
+      const url = window.URL.createObjectURL(new Blob([response.data]));
+      const link = document.createElement('a');
+      link.href = url;
+      link.setAttribute('download', feature.properties.name+'.pdf');
+      document.body.appendChild(link);
+      link.click();
+    });
+  }
